- 完善插件安装过程和菜单配置 - 在安装插件时增加依赖检测和自动安装 - 添加插件卸载功能 - 优化菜单配置，增加系统菜单标识 - 新增字典和定时任务安装功能 - 重构插件安装、更新和卸载的接口和逻辑

+++ app/admin/controller/AdminRuleController.php
@@ -265,6 +265,9 @@ class AdminRuleController extends XbController
             return $this->success('修改成功');
         }
         $data           = $model->toArray();
+        if ($data['is_system'] == '20') {
+            return $this->fail('系统菜单，禁止操作');
+        }
         $data['remote'] = $data['params'];
         $builder        = $this->formView();
         $builder->setMethod('PUT');
@@ -288,6 +291,9 @@ class AdminRuleController extends XbController
         // 批量删除
         $data = AdminRule::where('id', 'in', $ids)->select();
         foreach ($data as $model) {
+            if ($model['is_system'] == '20') {
+                return $this->fail('系统菜单，禁止操作');
+            }
             if (!$model->delete()) {
                 throw new Exception("ID:{$model['id']} 删除失败");
             }
